Create Bag ESP.lua
This is bag esp script. Feel free to try to understand whats there ;)
I've been using it for 6 years straight and never got banned for it,
i can make a conclusion that it is undetectable. 
Types of bags that it tracks:
Limited (doesnt exist in original game)
Christmas
Halloween
Resplendent
Ultimate
Legendary
Epic
Rare
Uncommon
Common